move fast_idiv to transcendental
